todo_turso/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==1.0.0
anyio==4.4.0
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docutils==0.21.2
fastcore==1.7.1
fastlite==0.0.8
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
libsql-client==0.3.1
libsql-experimental==0.0.40
MarkupSafe==2.1.5
multidict==6.1.0
oauthlib==3.2.2
packaging==24.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.4.2
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.0.2
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
sqlalchemy-libsql==0.1.0
sqlite_minutils==3.36.0.post4
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==12.0
wheel==0.44.0
yarl==1.11.1
